requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
whitenoise==6.9.0
